Moved contract loader to separate upload_contracts module
