+++ .travis.yml
@@ -6,27 +6,10 @@ notifications:
 python:
   - "2.7_with_system_site_packages"
 
-branches:
-  only:
-    - scaling
-
-addons:
-  apt:
-    packages:
-    - libmpc-dev
-    - libmpfr-dev
-    - libmpfr4
-    - libgmp-dev
-    - libblas-dev
-    - liblapack-dev
-    - gfortran
-    - python-numpy
-    - python-scipy
-
 install:
   - pip install -U pip
-  - travis_wait travis_retry pip install -r requirements.txt
-  - travis_retry pip install -r test_requirements.txt
+  - pip install -r requirements.txt
+  - pip install -r test_requirements.txt
 
 script:
   - py.test tests/trading_tests.py
