+++ test_requirements.txt
@@ -1,0 +1,2 @@
+iocapture>=0.1.2
+pytest
