+++ upload_contracts/__init__.py
@@ -1,0 +1,1 @@
+from .upload_contracts import ContractLoader
